Also check monocle for fullscreen

+++ yambar/.config/yambar/hyprland.sh
@@ -32,7 +32,7 @@ socat -U - "UNIX-CONNECT:$sock" | while read -r line; do
     monocle=$([ ${window_info[6]} -eq 1 ] && echo "true" || echo "false")
 
     # Toggle gammastep on fullscreen change
-    if [ "$fullscreen" = true ] && [ "$was_fullscreen" = false ]; then
+    if [ "$fullscreen" = true ] && [ "$monocle" = false ] && [ "$was_fullscreen" = false ]; then
         was_fullscreen=true
         pkill -USR1 gammastep
     elif [ "$fullscreen" = false ] && [ "$was_fullscreen" = true ]; then
